Rewrite protection
Lots of refactoring and completely changed the
way banning works.

Banning for short intervals did not work well
because there was no way to inform the player
for how long they were banned. Instead, we now
suspend shout and interact privileges for a time
and the player is informed for how long. If the
player reaches the banlevel 4 times, they are
permanently banned as before.

The suspension times increase exponentially;
3, then 9, then 27 minutes before banning.